+++ public/cpp/logic/public/public/mail/actor_mail.cpp
@@ -150,7 +150,7 @@ namespace ngl
 			);
 		}
 
-		if (handle_cmd::function(loperator, adata.get_data()->identifier(), lojson) == false)
+		if (handle_cmd::function(loperator, (int)adata.get_data()->identifier(), lojson) == false)
 		{
 			log_error()->print("GM actor_mail operator[{}] ERROR", loperator);
 		}
