+++ public/cpp/logic/public/public/ranklist/actor_ranklist.cpp
@@ -75,7 +75,7 @@ namespace ngl
 			);
 		}
 
-		if (handle_cmd::function(loperator, adata.get_data()->identifier(), lojson) == false)
+		if (handle_cmd::function(loperator, (int32_t)adata.get_data()->identifier(), lojson) == false)
 		{
 			log_error()->print("GM actor_mail operator[{}] ERROR", loperator);
 		}
